Splits vendor en extern-vendor, om de publieke stek wat lichter te maken
