+++ resources/assets/js/datatable.js
@@ -1,12 +1,9 @@
-// Wordt na app.js geladen.
-
 import {knopPost} from './knop';
 
 import $ from 'jquery';
 import initContext from './context';
 // Excel button in datatables.net-buttons/js/buttons.html5 checkt voor JSZip in window.
 import JSZip from 'jszip';
-window.JSZip = JSZip;
 
 /**
  * Knoop alle datatable plugins aan jquery.
@@ -27,6 +24,8 @@ import 'datatables.net-scroller';
 import 'datatables.net-select';
 import './lib/dataTables.childRow';
 import './lib/dataTables.columnGroup';
+
+window.JSZip = JSZip;
 
 /*!
  * csrdelft.dataTables.js
@@ -300,47 +299,3 @@ window.fnAutoScroll = function(tableId) {
         }
     }
 };
-
-/**
- * @see datatable.js
- * @see view/formulier/datatable/DataTable.php
- * @param tableId
- * @param response
- */
-export function fnUpdateDataTable(tableId, response) {
-    let $table = $(tableId);
-    let table = $table.DataTable();
-    // update or remove existing rows or add new rows
-    response.data.forEach((row) => {
-        let $tr = $('tr[data-uuid="' + row.UUID + '"]');
-        if ($tr.length === 1) {
-            if ('remove' in row) {
-                table.row($tr).remove();
-            }
-            else {
-                table.row($tr).data(row);
-                initContext($tr);
-            }
-        }
-        else if ($tr.length === 0) {
-            table.row.add(row);
-        }
-        else {
-            alert($tr.length);
-        }
-    });
-    table.draw(false);
-}
-
-/**
- * @see csrdelft.js
- * @param tableId
- * @returns {Array}
- */
-export function fnGetSelection(tableId) {
-    let selection = [];
-    $(tableId + ' tbody tr.selected').each(function () {
-        selection.push($(this).attr('data-uuid'));
-    });
-    return selection;
-}
